Move utils to hooks
